Update audit text and improve resource handling
This updates the audit failure text to only reference the number
of upgradeable resources. It also simplifies how resources are
tracked now that we don't output details about non-upgradeable
resources. It also only counts upgradeable resources that were
initiated by the main page for the count and score, so that the
count of upgradeable resources in the text matches the number of
items in the audit report table.